+++ components/select/option.vue
@@ -1,36 +1,18 @@
-<template>
-    <div
-        v-show="isShow"
-        ref="optionRef"
-        :class="classList"
-        @click="handleClick"
-    >
-        <slot :isChecked="isChecked">
-            <Ellipsis :triggerClass="`${prefixCls}-label`">
-                {{ label }}
-            </Ellipsis>
-            <CheckOutlined
-                v-if="isChecked"
-                :class="`${prefixCls}-checked-icon`"
-            />
-        </slot>
-    </div>
-</template>
 <script>
-import { computed, inject, onMounted, ref } from 'vue';
-import Ellipsis from '../ellipsis';
+import {
+    computed,
+    getCurrentInstance,
+    inject,
+    onBeforeMount,
+    onBeforeUnmount,
+    toRefs,
+} from 'vue';
+import { isArray, isString } from 'lodash-es';
 import { key } from './const';
-import getPrefixCls from '../_util/getPrefixCls';
-import CheckOutlined from '../icon/CheckOutlined';
-
-const prefixCls = getPrefixCls('select-option');
+import { getOptionId } from './helper';
 
 export default {
     name: 'FOption',
-    components: {
-        CheckOutlined,
-        Ellipsis,
-    },
     props: {
         value: {
             type: [String, Number, Boolean, Object],
@@ -45,54 +27,46 @@ export default {
             default: false,
         },
     },
-    setup(props) {
+    setup(props, ctx) {
         const parent = inject(key, null);
         if (!parent) {
             return console.error(
                 '[FOption]: FOption 必须搭配 FSelect 组件使用！',
             );
         }
-        const { isSelect, onSelect, addOption, filterText } = parent;
+        const instance = getCurrentInstance();
+        const optionId = getOptionId();
+        instance.optionId = optionId;
+        const { addOption, removeOption } = parent;
 
-        const isShow = computed(
-            () => !filterText.value || props.label.includes(filterText.value),
-        );
-
-        const isChecked = computed(() => isSelect(props.value));
-
-        const classList = computed(() => {
-            const arr = [
-                prefixCls,
-                isChecked.value && 'is-checked',
-                props.disabled && 'is-disabled',
-            ];
-            return arr.filter(Boolean);
-        });
-
-        const optionRef = ref();
-
-        onMounted(() => {
-            const text = optionRef.value.innerText;
-            const option = { ...props };
-            if (!option.label && text) {
-                option.label = text;
+        // 当插槽只是string时，通过slot计算label
+        let label = '';
+        if (!props.label) {
+            const vNodes = ctx.slots.default();
+            if (
+                isArray(vNodes) &&
+                vNodes.length === 1 &&
+                isString(vNodes[0].children)
+            ) {
+                label = vNodes[0].children;
             }
+        }
+
+        onBeforeMount(() => {
+            const option = {
+                id: optionId,
+                ...toRefs(props),
+                ctx,
+            };
+            option.label = computed(() => label || props.label);
             addOption(option);
         });
 
-        const handleClick = () => {
-            if (props.disabled) return;
-            onSelect(props.value);
-        };
+        onBeforeUnmount(() => {
+            removeOption(optionId);
+        });
 
-        return {
-            isChecked,
-            prefixCls,
-            handleClick,
-            classList,
-            isShow,
-            optionRef,
-        };
+        return () => null;
     },
 };
 </script>