+++ components/version/index.js
@@ -1,3 +1,3 @@
-const version = '0.1.6';
+const version = '0.1.8';
 
 export default version;